WIP: DB connection resource
The goal is to have a bevy `Resource`, which I can use to query the ~org-roam~
database. I switched the async runtime for ~sqlx~ to ~async_std~ to better align
with what bevy is using. With the ~async_std::task::block_on~ function I manage
to get everything to compile, but I have not yet tested if it works.